+++ com.unity.netcode.adapter.utp/CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this package will be documented in this file. The format 
 
 ### Added
 
+- New parameters are available to simulate network conditions (delay, jitter, packet loss) in the editor and in development builds. The parameters are available under the 'Debug Simulator' section of the 'Unity Transport' component, or can be set with the `SetDebugSimulatorParameters` call. (#1745)
+
 ### Changed
 
 - Updated Unity Transport package to 1.0.0-pre.13. (#1696)